move ticket model logic into query.js and add constants.js
