+++ assets/src/data/model/ticket/constants.js
@@ -1,0 +1,1 @@
+export const MODEL_NAME = 'ticket';
